# Remove LINQ usage for IEnumerable comparisons