--- a/src/OpenApi/src/Comparers/OpenApiAnyComparer.cs
+++ b/src/OpenApi/src/Comparers/OpenApiAnyComparer.cs
@@ -3,10 +3,11 @@
 
 using System.Linq;
 using Microsoft.OpenApi.Any;
+using Microsoft.OpenApi.Interfaces;
 
 namespace Microsoft.AspNetCore.OpenApi;
 
-internal sealed class OpenApiAnyComparer : IEqualityComparer<IOpenApiAny>
+internal sealed class OpenApiAnyComparer : IEqualityComparer<IOpenApiAny>, IEqualityComparer<IOpenApiExtension>
 {
     public static OpenApiAnyComparer Instance { get; } = new OpenApiAnyComparer();
 
@@ -29,8 +30,8 @@ internal sealed class OpenApiAnyComparer : IEqualityComparer<IOpenApiAny>
             (x switch
             {
                 OpenApiNull _ => y is OpenApiNull,
-                OpenApiArray arrayX => y is OpenApiArray arrayY && arrayX.SequenceEqual(arrayY, Instance),
-                OpenApiObject objectX => y is OpenApiObject objectY && objectX.Keys.Count == objectY.Keys.Count && objectX.Keys.All(key => objectY.TryGetValue(key, out var yValue) && Equals(objectX[key], yValue)),
+                OpenApiArray arrayX => y is OpenApiArray arrayY && ComparerHelpers.ListEquals(arrayX, arrayY, Instance),
+                OpenApiObject objectX => y is OpenApiObject objectY && ComparerHelpers.DictionaryEquals(objectX, objectY, Instance),
                 OpenApiBinary binaryX => y is OpenApiBinary binaryY && binaryX.Value.SequenceEqual(binaryY.Value),
                 OpenApiInteger integerX => y is OpenApiInteger integerY && integerX.Value == integerY.Value,
                 OpenApiLong longX => y is OpenApiLong longY && longX.Value == longY.Value,
@@ -77,5 +78,40 @@ internal sealed class OpenApiAnyComparer : IEqualityComparer<IOpenApiAny>
         });
 
         return hashCode.ToHashCode();
+    }
+
+    public bool Equals(IOpenApiExtension? x, IOpenApiExtension? y)
+    {
+        if (x is null && y is null)
+        {
+            return true;
+        }
+
+        if (x is null || y is null)
+        {
+            return false;
+        }
+
+        if (object.ReferenceEquals(x, y))
+        {
+            return true;
+        }
+
+        if (x is IOpenApiAny openApiAnyX && y is IOpenApiAny openApiAnyY)
+        {
+            return Equals(openApiAnyX, openApiAnyY);
+        }
+
+        return x.Equals(y);
+    }
+
+    public int GetHashCode(IOpenApiExtension obj)
+    {
+        if (obj is IOpenApiAny any)
+        {
+            return GetHashCode(any);
+        }
+
+        return obj.GetHashCode();
     }
 }