--- a/src/OpenApi/src/Comparers/OpenApiExternalDocsComparer.cs
+++ b/src/OpenApi/src/Comparers/OpenApiExternalDocsComparer.cs
@@ -1,7 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System.Linq;
+using Microsoft.OpenApi.Interfaces;
 using Microsoft.OpenApi.Models;
 
 namespace Microsoft.AspNetCore.OpenApi;
@@ -27,8 +27,7 @@ internal sealed class OpenApiExternalDocsComparer : IEqualityComparer<OpenApiExt
 
         return x.Description == y.Description &&
             x.Url == y.Url &&
-            x.Extensions.Count == y.Extensions.Count
-            && x.Extensions.Keys.All(k => y.Extensions.TryGetValue(k, out var yValue) && yValue == x.Extensions[k]);
+            ComparerHelpers.DictionaryEquals((Dictionary<string, IOpenApiExtension>)x.Extensions, (Dictionary<string, IOpenApiExtension>)y.Extensions, OpenApiAnyComparer.Instance);
     }
 
     public int GetHashCode(OpenApiExternalDocs obj)