Contributes towards https://github.com/dotnet/aspnetcore/issues/56829. Refresh of https://github.com/dotnet/aspnetcore/pull/56599.

- Replace use of `Enumerable.All` with a custom implementation of `DictionaryEquals` and `ListEquals` to avoid overhead due to lambda capture in existing comparers.
- Cast to concrete collection types in `DictionaryEquals` and `ListEquals` to avoid interface dispatch overhead. See https://github.com/dotnet/aspnetcore/pull/56599#discussion_r1671411261 fore more info.

Resulting benchmarks from this change are as follows:

**Before**

| Method           | EndpointCount | Mean        | Error     | StdDev    | Op/s     | Gen0     | Gen1     | Allocated  |
|----------------- |-------------- |------------:|----------:|----------:|---------:|---------:|---------:|-----------:|
| GenerateDocument | 10            |    303.7 μs |   2.43 μs |   2.27 μs | 3,292.87 |   3.9063 |        - |  538.06 KB |
| GenerateDocument | 100           |  2,843.4 μs |  35.50 μs |  29.65 μs |   351.69 |  31.2500 |  15.6250 |    4891 KB |
| GenerateDocument | 1000          | 31,125.8 μs | 492.27 μs | 460.47 μs |    32.13 | 333.3333 | 166.6667 | 48417.5 KB |

**After**

| Method           | EndpointCount | Mean        | Error     | StdDev    | Op/s     | Gen0     | Allocated   |
|----------------- |-------------- |------------:|----------:|----------:|---------:|---------:|------------:|
| GenerateDocument | 10            |    270.5 us |   1.39 us |   1.30 us | 3,697.22 |   2.9297 |    434.9 KB |
| GenerateDocument | 100           |  2,484.8 us |  25.84 us |  24.17 us |   402.45 |  31.2500 |  3944.06 KB |
| GenerateDocument | 1000          | 27,413.2 us | 219.91 us | 194.95 us |    36.48 | 166.6667 | 39030.74 KB |